% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efreadr.R
\docType{package}
\name{efreadr-package}
\alias{efreadr-package}
\title{Read European Fluxes CSV Files}
\description{
The European Eddy Fluxes Database Cluster distributes fluxes of different Green House Gases measured mainly using the eddy covariance technique acquired in sites involved in EU projects but also single sites in Europe, Africa and others continents that decided to share their measurements in the database (cit. http://gaia.agraria.unitus.it ). The package provides two functions to load and row-wise bind CSV files distributed by the database. Currently only L3 and L4 (L=Level), half-hourly and daily (aggregation) files are supported.
}
\author{
Marco Bascietto \email{marco.bascietto@crea.gov.it}
}
\references{
Source code is hosted at GitHub (\url{https://github.com/mbask/efreadr})
}
\keyword{package}

